Added methods to return their email. I also added the add a number to their file. I also added a method to reset the register number to 0
